Bootstrap Cribl Log Search Application
The purpose of this change is to lay the foundation for the Cribl
log search application.

As part of this change, the Spring framework has been setup, the
initial version of the search page has been defined and controllers
have been created.

In the subsequent changes, we'll continue to add additional logic
to the files in this commit.

Testing Done:
A verification has been done to see if the application (server)
comes up fine and the page gets rendered on the browser as expected.

Command Ran:
./gradlew tomcatRun --no-daemon -debug

Browser URL - http://localhost:8080/