Add initial support for streaming response
This change adds support for streaming, both on the client
and the server side.

On the client side, the client will display the logs
incrementally (periodically) by checking if any updates have
been sent by the server.

On the server side, the controller supports returning a
streaming body as the response, with the ability to
immediately ack back to the client, while keeping the connection
open to flush further updates.

Testing:

Verified if the client (browser) incrementally displays a sample
(dummy) data sent by the server.