Query logs based on filters
This change adds support for querying the logs, based on the
filters specified by the client.

The filters can be either the filename, match count or a pattern or
any combination of these three.

The retrieved logs are streamed back to the caller (client).

Testing Done:
Verified if the logs are streamed based on filters via the UI
(browser).

Also, added unit tests to verify the changes.